add docker compose yaml
This is the YAML that defines for docker compose what resources it will need and how to configure them.
This YAML will contain everything we need, which as of now is a MySQL container for containing user information - usernames, passwords, and other associated metadata like highscores and such. Python has ways of querying this data which will allow us to use it in our programs. Additionally we have another container named 'python' which will run our python code.
We can have other containers if we need. There are containers for just about everything nowadays. NGINX server? We can do it. JavaScript server? Yep. Etc. Whatever we need it can be done.
In the future we will probably have a Jenkins server that we will be able to use to easily deploy the containers we create.